removed useless update to chunk_review record
This was calling the update twice for each setTransaltion without
adding anything useful.

The count of qa_entries is not needed here because qa_entries do not
change during setTranslation. The comment about propagation was also not
useful in understanding the purpose of these two lines of code.

It's very likely a leftover of a previous change when the update was not
being performed on earlier methods.